django/library/aws_libs.py
"""
This file was created at Smartbuzz Inc.
For more information visit http://www.smartbuzzinc.com
"""
from storages.backends.s3boto3 import S3Boto3Storage


class MediaStorage(S3Boto3Storage):
	location = 'media'
	file_overwrite = False