django/requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
aws-xray-sdk==0.96
backports.csv==1.0.7
bcrypt==3.1.4
beautifulsoup4==4.7.1
billiard==3.5.0.3
boto==2.48.0
boto3==1.9.206
boto3-python3==1.9.139
botocore>=1.12.206
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.1
coreschema==0.0.4
coverage==4.4.2
cryptography==2.2.2
decorator==4.3.0
defusedxml==0.5.0
dicttoxml==1.7.4
diff-match-patch==20181111
Django==1.11.7
django-braces==1.11.0
django-cors-headers==2.1.0
django-import-export==1.0.1
django-jet==1.0.8
django-mptt==0.8.7
django-multiupload==0.5.2
django-nose==1.4.5
django-oauth-toolkit==1.0.0
django-polymorphic==1.3
django-redis==4.9.0
django-resized==0.3.9
django-rest-framework-social-oauth2==1.0.8
django-rest-swagger==2.1.2
django-ses==0.8.5
django-storages==1.6.5
django-suit==0.2.26
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
et-xmlfile==1.0.1
Fabric==1.14.0
Fabric3==1.13.1.post1
feedparser==5.2.1
future==0.16.0
google-api-python-client==1.6.4
gunicorn==19.7.1
hashids==1.2.0
HTMLParser==0.0.2
httplib2==0.10.3
idna==2.6
imageio==2.3.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.9.6
jmespath==0.9.3
jsonpickle==0.9.6
kombu==4.1.0
mailchimp3==3.0.1
MarkupSafe==1.0
moviepy==0.2.3.4
nose==1.3.7
numpy==1.14.3
oauth2client==4.1.2
oauthlib==2.0.4
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.2
paramiko==2.4.1
Pillow==5.1.0
psycopg2==2.7.3.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyJWT==1.5.3
pymongo==3.6.1
PyNaCl==1.2.1
python-dateutil==2.6.1
python-dotenv==0.10.0
python3-openid==3.1.0
pytz==2017.2
PyYAML==5.1
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer>=0.2.0
serializers==0.2.4
simplejson==3.11.1
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
soupsieve==1.9
sqlparse==0.2.4
tablib==0.13.0
tqdm==4.23.4
unicode==2.7
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wrapt==1.10.11
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.11.0
